feat: Add build and packaging and modularize
