+++ requirements.txt
@@ -1,0 +1,1 @@
+setuptools~=66.1.1
