+++ run.py
@@ -1,0 +1,6 @@
+#! /usr/bin/env python
+
+from spyglass.cli import main
+
+if __name__ == "__main__":
+    raise SystemExit(main())
