+++ setup.cfg
@@ -1,0 +1,17 @@
+[metadata]
+name = spyglass
+version = 0.0.1
+description = A simple mjpeg server for Picamera2
+url = https://github.com/roamingthings/spyglass
+license = Apache License 2.0
+license_files =
+   LICENSE
+   NOTICE
+
+[options]
+packages = find:
+python_requires = >=3.8.1
+
+[options.entry_points]
+console_scripts =
+    spyglass = spyglass.cli:main
