+++ spyglass/__main__.py
@@ -1,0 +1,7 @@
+"""Module allowing for ``python -m spyglass ...``."""
+from __future__ import annotations
+
+from spyglass.cli import main
+
+if __name__ == "__main__":
+    raise SystemExit(main())
